langchain/pyproject.toml
[project]
name = "popquiz-langchain"
version = "0.1.0"
description = "LangChain AI microservice for PopQuiz — topic-based and RAG quiz generation using Google Gemini"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-chroma>=0.1.0",
    "langchain-google-genai>=2.0.0",
    "chromadb>=0.5.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "Pillow>=10.0.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.3.0",
]

[tool.uv]
package = false
dev-dependencies = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
